modules/observability/ai_alerts/requirements.txt
flask
google-cloud-aiplatform
requests
vertexai
google-genai
